[TASK] Replace csConvObj calls
The CharsetConverter object is agnostic
of charset and environment, and a singleton on top.

There are still several calls in the TYPO3 core
that tries to fetch csConv from TSFE or LANG
which relies heavily on the environment.

The calls are replaced with simple instance
calls where possible.

Resolves: #73441
Releases: master
Change-Id: I098f17b1cdbcc1ebbbb7359cfec518c93597aad4
Reviewed-on: https://review.typo3.org/46627
Reviewed-by: Stefan Froemken <froemken@gmail.com>
Tested-by: Stefan Froemken <froemken@gmail.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>