[!!!][TASK] Remove example image in TS constants description
In TYPO3 it was possible to show an image for a category in
the Constants Editor, if the constants editor had configured a category
like this (also defined in the Constants Editor):

TSConstantEditor.basic.image = EXT:sys_note/ext_icon.png

As the functionality is hardly used and better options for styling
and displaying images e.g. via user functions is possible, the
image option has been removed.

With that, the following breaking changes in the related PHP code
happened:
- Removed ExtendedTemplateService::$ext_localGfxPrefix
- Removed ExtendedTemplateService::$ext_localWebGfxPrefix
- Removed ExtendedTemplateService::ext_getTSCE_config_image
- Removed ConfigurationForm::ext_getTSCE_config_image
- Dropped parameter 2+3 of ConfigurationForm::ext_initTSstyleConfig

Resolves: #77062
Releases: master
Change-Id: Ic7d89aeaa4f236a92a07dd2fe0a25dbe026825cb
Reviewed-on: https://review.typo3.org/48922
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Thomas Schlumberger <thomas@b13.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>