[FEATURE] Display (un)substituted TS constant
Display value of TS constants on "UN-substituted constants in green"
mode and constant name when on "Substituted constant in green" mode.

Before this patch toggling between both modes was necessary to simply
get an idea of what the value/variable name was while being on the
'other' mode.

As a side effect this patch improves the search feature of the object
browser when searching for a constant value or name, i.e. 'maxWInText'
while being on "Substituted" mode.

Resolves: #40729
Releases: master
Change-Id: Ic8c3f1b96b3b440a3ed4d098604e29d018129451
Reviewed-on: https://review.typo3.org/50190
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>