[TASK] Improve TS Template module information on root level list
The template tool on root level, which displays an overview of pages
including templates, now displays instead of the count of templates
all related templates of a page with several informations like:
- title
- state (by icon)
- is root or extend
- direct link to edit (with ID ans site-title as link title)
Pages in rootline, which do not includes templates itself are no longer
linked to the template modul.
Database query has been streamlined, and all render logic of
Html specific output has been moved to FLUID templating.
This offers the possibility to add filter or search to this view.
An issue with page sorting in page tree, when rootline pages does not
contain TS templates has been fixed.

Resolves: #81430
Releases: master, 8.7
Change-Id: I5250c53df1acc4a1a29839ce4a472f3e863ae992
Reviewed-on: https://review.typo3.org/57856
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Mathias Brodala <mbrodala@pagemachine.de>
Tested-by: Mathias Brodala <mbrodala@pagemachine.de>