[BUGFIX] Resolve workspace records for template module overview
BackendWorkspaceRestriction automatically resolves specific records
already for a given workspace. In most cases this seems to be fine to
omit the usual workspace overlay processing. However, especially when
dealing with references and page tree assignments, it just contains
wrong information.

For instance all workspace versions have the `pid` column set to -1
which of course cannot be resolved as valid page then.

Resolves: #85917
Releases: master, 8.7
Change-Id: Ibe20037b25770bea8f58d1a8bba785ca43ac26cb
Reviewed-on: https://review.typo3.org/57982
Tested-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>