[BUGFIX] React on delete actions via context menu
If a record gets deleted via context menu no further handling could be
added as the promise returned by AjaxDataHandler is available in an
encapsulated scope only.

To bypass this issue, the AjaxDataHandler now optionally emits events to
listen on. This patch implements event listeners for the record list and
the information module of "Template tools".

Resolves: #90791
Releases: master
Change-Id: I0b659e695f31b2f3ab40adc17421f8b036d5aab6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63811
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>