[BUGFIX] Streamline AjaxDataHandler event names
The events dispatched by AjaxDataHandler are changed to be always named
`typo3:datahandler:process`. This is done to ease registering event
listeners as the executed action is part of the sent payload already.

Since this feature is not documented and propagated yet, this is
considered being not breaking.

Resolves: #92028
Releases: master, 10.4
Change-Id: Ibe69d7297bbc6e17fd744ebf96c3f88f719e533d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/65418
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>