[BUGFIX] Do not emit redirect() via PHP but via PSR-7
In order to allow further manipulation of Response objects,
using "throw new ImmediateResponseException" is much
more flexible than calling HttpUtility::redirect() which immediately
stops the PHP process. This way, it is not even possible
for developers to manipulate the response, or even
test this behavior in functional tests with TYPO3 Core.

The change removes all usages of HttpUtility::redirect().
The method will be deprecated at a later stage.

Resolves: #93004
Releases: master
Change-Id: I9bd0db2b2ee0c15b39b38168d67e6d78ba4be2db
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/67038
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Daniel Gorges <daniel.gorges@b13.com>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Daniel Gorges <daniel.gorges@b13.com>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>