[FEATURE] Revamp TypoScript Template Analyzer
This establishes the first usage of the new TypoScript
parser and leverages some of its improved functionality:
A rewrite of the template analyzer backend module based
on the new parser gives a series of new features.

Most importantly, the TypoScript include tree now
resolves all @import and INCLUDE_TYPOSCRIPT includes
and shows them as single nodes within the include tree.

Additionally, the template analyzer is now condition
aware: It finds all conditions within the include
tree and allows to simulate include tree restrictions
based on condition verdicts.

The major outcome of this is that the analyzer now
shows in-depth includes and allows much more fine
grained analysis of all includes that play a role
on a single page.

Known issue: t3editor (codemirror) sometimes fails to
initialize properly to render TypoScript source when
clicking a node in the include tree. This is probably
some JavaScript load-order race condition and will be
tackled with a dedicated patch.

Change-Id: I2761e1dfc9e511904b73d48de41198b84aee6447
Resolves: #97941
Related: #97816
Releases: main
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75178
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>