[BUGFIX] Correct treelist-renderings
This is the first step to correct the CSS list-tree. Even it is
documented in the style guide extension it was managed to have
several kinds of implementation through the core.

To make it clear and force people to look up the actual use, we are
now also streamlining the naming scheme.

Old: .list-tree
New: .treelist

This led to a misplacement of the open/collapsed indicator/toggle
of tree branches. The control element was somehow moved inside the
treelist-group, which is purely intended to handle a single row of
informational content. In addition collapse classes were mixed up
with tree stylings, which lead to cut-off indicators, and sluggish
opening and close animations.

Helpers to make tables responsive where used out of place to achieve
similar but different behavior for panels that are intended to
overflow if the content is too long.

The diff component was used in an unintended space and was not
used as documented. That made manual adaptions necessary that now
have been revoked. Instead, we now only apply the table structure
only if the nesting is correct. So the parts of the component can
be used in other places for now. While we are here, we are adapting
and streamlining the colors of the diff component.

All randomly placed text-monospace classes were removed from the
table rendering and are now only applied to the information group
if necessary. This avoids having badges, buttons, etc restyled
to a monospace font.

The styling adaptions in the import/export module were incomplete
and several necessary corrections were missing. We now made it
easier to define the background color for the element.

Misplaced control CSS classes that seem to be leftovers have been
removed from elements they don`t belong to.

As mentioned, this patch is part of the ongoing cleanup and has no
intention to be finished or solve all the remaining issues.

Resolves: #100476
Releases: main
Change-Id: I675496fbcc1205ce243daa74ff99ade28fb03efe
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78485
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>