[TASK] Warn about empty @import in TypoScript
When @import or <INCLUDE_TYPOSCRIPT: lines do not
resolve to any include, this probably indicates
a broken import statement.

To help integrators find and fix those, the
Page TSconfig and TypoScript backend modules now
list them as "Syntax scanner warnings".

We're adding a functional test to verify the detection
works. This will ensure it does not break when we refactor
some details of the include tree structure later. Doing as
described in #102102 and #102103 would simplify and
streamline various details, but it's currently unclear if
we should still do this in v12.

Resolves: #102064
Related: #100218
Related: #97816
Related: #102102
Related: #102103
Releases: main, 12.4
Change-Id: I1bba0cead97ff81b38c97f18581275edee01ea63
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81352
Tested-by: core-ci <typo3@b13.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Markus Klein <markus.klein@typo3.org>