[TASK] Simplify TypoScript -> Included TypoScript view
When we redesigned the TypoScript main module in v12, we
thought it is a good idea to make the "Included TypoScript"
condition aware, similar to "Active TypoScript": @import
statements within condition bodies were reflected in the
rendered include tree and condition verdicts could be
toggled, simulating their effect on the resulting include
tree.

This however broke in v12 quickly after first
implementation patches, leaving only a "matched" or
"not matched" badge per listed condition in the tree,
without further useful impact.

Note conditions are rendered as "title" in the tree since
v13 #105276, which helps the overal usefulness of the
"Include tree" a lot.

The patch removes the entire condition list and verdict
selection panel in "Included TypoScript" since it does
not contain good information nor functionality in this
view.

Resolves: #105276
Related: #103990
Related: #97816
Releases: main
Change-Id: I57f0ee3117d39f01f839843541148881d7cbf810
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/86559
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>