[TASK] Merge backend modules "Setup" and "Settings"
A new "Tiles" view option has been added to the overview of the
"Sites" >"Setup" backend module. This new view provides users with a
quick overview of the sites in the installation. More detailed
information about a site, such as the settings configured in the site
configuration and active site settings, can be viewed in a newly
introduced detail view.

Sites without a site configuration are now displayed in a separate
section, similar to the unassigned site configurations.

Finally, a small information section has been added to the overview,
indicating how many sites are in the installation, how many of them
lack a site configuration, and the number of unassigned site
configurations.

Furthermore, the overview page of the backend module
"Sites" > "Settings" was removed and the existing additional views
of that module were moved to the module "Sites" > "Setup".

Resolves: #108265
Releases: main
Change-Id: I3776b9effd5595d62f423374de637017b3132b02
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91359
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Jasmina Ließmann <typo3@minapok.eu>
Reviewed-by: Jasmina Ließmann <typo3@minapok.eu>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Georg Ringer <georg.ringer@gmail.com>